+++ src/terms/terms.service.ts
@@ -1,10 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateTermDto } from './dto/create-term.dto';
 import { UpdateTermDto } from './dto/update-term.dto';
+import { ApiErrorResponseBuilder } from 'src/common/api-error-builder';
+import { ErrorCode } from 'src/common/error-codes';
 
 @Injectable()
 export class TermsService {
+  private readonly logger = new Logger(TermsService.name);
   constructor(private prisma: PrismaService) {}
 
   async create(createTermDto: CreateTermDto) {
@@ -57,12 +60,22 @@ export class TermsService {
   }
 
   async remove(id: string) {
+    // check if the term is used in course sections
+    const courseSections = await this.prisma.courseSection.findMany({
+      where: { termId: id },
+    });
+    if (courseSections.length > 0) {
+      const errorResponse = ApiErrorResponseBuilder.create(
+        ErrorCode.TRMA,
+        'Term is used in course sections'
+      )
+        .withLogger(this.logger)
+        .build();
+      throw new BadRequestException(errorResponse);
+    }
     try {
       return await this.prisma.term.delete({
         where: { id },
-        include: {
-          schoolYear: true,
-        },
       });
     } catch (error) {
       throw new NotFoundException(`Term with ID ${id} not found`);
